requirements.txt
# ━━━ BASE DE DATOS ━━━
mysql-connector-python==8.0.33
SQLAlchemy==2.0.23

# ━━━ MACHINE LEARNING ━━━
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.11.4

# ━━━ PROCESAMIENTO DE DATOS ━━━
pandas==2.0.3
openpyxl==3.1.2

# ━━━ PROCESAMIENTO DE TEXTO ━━━
nltk==3.8.1
unidecode==1.3.6

# ━━━ API REST (opcional para panel admin) ━━━
flask==3.0.0
flask-cors==4.0.0

# ━━━ VOZ ━━━
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14

# ━━━ IMPRESIÓN ━━━
python-escpos==3.0
pyusb==1.2.1
Pillow==10.1.0

# ━━━ PDF (BACKUP) ━━━
reportlab==4.0.7

# ━━━ UTILIDADES ━━━
python-dotenv==1.0.0
python-dateutil==2.8.2